tests: refractor tests: conftest and mocks
